Refactor DefaultBlockWorker
